# openapi.yaml
openapi: 3.0.3
info:
  title: Recruitment Intelligence API
  version: 1.0.0
paths:
  /api/auth/login/:
    post:
      operationId: auth_login_create
      description: |-
        Takes a set of user credentials and returns an access and refresh JSON web
        token pair to prove the authentication of those credentials.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/VerifiedToken'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/VerifiedToken'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/VerifiedToken'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/VerifiedToken'
          description: ''
  /api/auth/logout/:
    post:
      operationId: auth_logout_create
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Logout'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Logout'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Logout'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
  /api/auth/refresh/:
    post:
      operationId: auth_refresh_create
      description: |-
        Takes a refresh type JSON web token and returns an access type JSON web
        token if the refresh token is valid.
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/TokenRefresh'
        required: true
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/TokenRefresh'
          description: ''
  /api/auth/register/:
    post:
      operationId: auth_register_create
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Register'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Register'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Register'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Register'
          description: ''
  /api/auth/verify/:
    post:
      operationId: auth_verify_create
      tags:
      - auth
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Verify'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Verify'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Verify'
        required: true
      security:
      - jwtAuth: []
      - {}
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '400':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
...
  /api/candidates/:
    get:
      operationId: candidates_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - candidates
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedCandidateList'
          description: ''
    post:
      operationId: candidates_create
      tags:
      - candidates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Candidate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Candidate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
          description: ''
  /api/candidates/{id}/:
    get:
      operationId: candidates_retrieve
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this candidate.
        required: true
      tags:
      - candidates
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
          description: ''
    put:
      operationId: candidates_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this candidate.
        required: true
      tags:
      - candidates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Candidate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Candidate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Candidate'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
          description: ''
    patch:
      operationId: candidates_partial_update
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this candidate.
        required: true
      tags:
      - candidates
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PatchedCandidate'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/PatchedCandidate'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/PatchedCandidate'
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Candidate'
          description: ''
    delete:
      operationId: candidates_destroy
      parameters:
      - in: path
        name: id
        schema:
          type: integer
        description: A unique integer value identifying this candidate.
        required: true
      tags:
      - candidates
      security:
      - jwtAuth: []
      responses:
        '204':
          description: No response body
  /api/interview-feedback/:
    get:
      operationId: interview_feedback_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - interview-feedback
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedFeedbackList'
          description: ''
    post:
      operationId: interview_feedback_create
      tags:
      - interview-feedback
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Feedback'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Feedback'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Feedback'
        required: true
      security:
      - jwtAuth: []
      responses:
        '201':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Feedback'
          description: ''
  /api/interview-slots/:
    get:
      operationId: interview_slots_list
      parameters:
      - name: page
        required: false
        in: query
        description: A page number within the paginated result set.
        schema:
          type: integer
      tags:
      - interview-slots
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PaginatedSlotList'
          description: ''
...
  /api/screening/predict/:
    post:
      operationId: screening_predict_create
      tags:
      - screening
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Predict'
          application/x-www-form-urlencoded:
            schema:
              $ref: '#/components/schemas/Predict'
          multipart/form-data:
            schema:
              $ref: '#/components/schemas/Predict'
        required: true
      security:
      - jwtAuth: []
      responses:
        '200':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
        '503':
          content:
            application/json:
              schema:
                type: object
                additionalProperties: {}
          description: ''
components:
  schemas:
...
    Candidate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        candidate_name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          maxLength: 20
        college:
          type: string
          maxLength: 200
        applied_role:
          type: integer
        batch:
          type: integer
          readOnly: true
          nullable: true
        skills:
          type: string
        experience_months:
          type: integer
          maximum: 600
          minimum: 0
        notice_period_days:
          type: integer
          maximum: 365
          minimum: 0
        expected_salary:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        resume_text:
          type: string
        portfolio_url:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        historical_selection_status:
          type: string
          maxLength: 15
        status:
          $ref: '#/components/schemas/CandidateStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
      required:
      - applied_role
      - batch
      - candidate_name
      - created_at
      - email
      - expected_salary
      - experience_months
      - id
      - notice_period_days
      - phone
      - skills
      - updated_at
    CandidateStatusEnum:
      enum:
      - NEW
      - SHORTLISTED
      - INTERVIEW
      - SELECTED
      - WAITLISTED
      - REJECTED
      type: string
      description: |-
        * `NEW` - New
        * `SHORTLISTED` - Shortlisted
        * `INTERVIEW` - Interview
        * `SELECTED` - Selected
        * `WAITLISTED` - Waitlisted
        * `REJECTED` - Rejected
...
    Logout:
      type: object
      properties:
        refresh:
          type: string
      required:
      - refresh
...
    PaginatedCandidateList:
      type: object
      required:
      - count
      - results
      properties:
        count:
          type: integer
          example: 123
        next:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=4
        previous:
          type: string
          nullable: true
          format: uri
          example: http://api.example.org/accounts/?page=2
        results:
          type: array
          items:
            $ref: '#/components/schemas/Candidate'
...
    PatchedCandidate:
      type: object
      properties:
        id:
          type: integer
          readOnly: true
        candidate_name:
          type: string
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        phone:
          type: string
          maxLength: 20
        college:
          type: string
          maxLength: 200
        applied_role:
          type: integer
        batch:
          type: integer
          readOnly: true
          nullable: true
        skills:
          type: string
        experience_months:
          type: integer
          maximum: 600
          minimum: 0
        notice_period_days:
          type: integer
          maximum: 365
          minimum: 0
        expected_salary:
          type: string
          format: decimal
          pattern: ^-?\d{0,10}(?:\.\d{0,2})?$
        resume_text:
          type: string
        portfolio_url:
          oneOf:
          - type: string
            format: uri
            maxLength: 200
          - type: string
            maxLength: 0
        historical_selection_status:
          type: string
          maxLength: 15
        status:
          $ref: '#/components/schemas/CandidateStatusEnum'
        created_at:
          type: string
          format: date-time
          readOnly: true
        updated_at:
          type: string
          format: date-time
          readOnly: true
...
    Register:
      type: object
      properties:
        username:
          type: string
          description: Required. 150 characters or fewer. Letters, digits and @/./+/-/_
            only.
          pattern: ^[\w.@+-]+$
          maxLength: 150
        email:
          type: string
          format: email
          maxLength: 254
        password:
          type: string
          writeOnly: true
          minLength: 8
      required:
      - email
      - password
      - username
...
    TokenRefresh:
      type: object
      properties:
        access:
          type: string
          readOnly: true
        refresh:
          type: string
      required:
      - access
      - refresh
...
    VerifiedToken:
      type: object
      properties:
        username:
          type: string
          writeOnly: true
        password:
          type: string
          writeOnly: true
      required:
      - password
      - username
    Verify:
      type: object
      properties:
        email:
          type: string
          format: email
        code:
          type: string
          pattern: ^\d{6}$
      required:
      - code
      - email
  securitySchemes:
    jwtAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT